feat: SSH remote transport, store crate, kiki:// scheme
Add ssh:// remote scheme for syncing over SSH without a daemon on the
server. The local daemon spawns 'ssh user@host kiki kk serve /path' and
speaks a length-prefixed protobuf protocol over stdin/stdout.

New store crate: extract RemoteStore trait, BlobKind, CasOutcome,
FsRemoteStore, and the framing protocol into a shared crate used by
both the daemon and CLI.

New CLI subcommand: 'kiki kk serve <path>' serves a FsRemoteStore over
stdin/stdout. No daemon needed on the server.

New daemon module: SshRemoteStore spawns SSH + serve, implements
RemoteStore over the pipe with lazy connection.

Rename grpc:// to kiki:// (grpc:// kept as alias).

Clippy cleanups: collapsed nested ifs, removed unnecessary cast,
removed unused constant.

+++ daemon/src/git_store.rs
@@ -831,7 +831,7 @@ fn signature_from_proto(
         email: proto.email,
         timestamp: jj_lib::backend::Timestamp {
             timestamp: jj_lib::backend::MillisSinceEpoch(ts.millis_since_epoch),
-            tz_offset: ts.tz_offset as i32,
+            tz_offset: ts.tz_offset,
         },
     })
 }
